"Changed Blog link to 'blogsite.html' and added feature cards"
This commit message describes the changes made in the code diff above. It mentions that the Blog link in the navigation menu has been updated to point to 'blogsite.html', and that new feature cards have been added to the main content section. The message is concise and accurately reflects the changes made in the code.